+++ span.go
@@ -47,7 +47,10 @@ ReferencesLoop:
 	for _, ref := range opts.Options.References {
 		switch ref.Type {
 		case ot.ChildOfRef, ot.FollowsFromRef:
-			refCtx := ref.ReferencedContext.(SpanContext)
+			refCtx, ok := ref.ReferencedContext.(SpanContext)
+			if !ok {
+				break ReferencesLoop
+			}
 			sp.raw.Context.TraceID = refCtx.TraceID
 			sp.raw.ParentSpanID = refCtx.SpanID